Version 2.0.3
Guild:
  ~ Improved description of command [CmdWelcome.java]

NSFW:
  ~ Added check for if own message is not null. Hopefully fixing an annoying bug [CmdFuck.java & CmdYFuck.java]

Message:
  ~ Added Regex to change @everyone and @here to everyone and here respectively.

+++ src/main/java/site/purrbot/bot/commands/nsfw/CmdFuck.java
@@ -148,7 +148,8 @@ public class CmdFuck implements Command{
                         ev -> {
                             if(ev.getReactionEmote().getName().equals("❌")){
                                 try{
-                                    message.delete().queue();
+                                    if(message != null) 
+                                        message.delete().queue();
                                 }catch(Exception ex){
                                     logger.warn(String.format(
                                             "Couldn't delete own message for CmdFuck. Reason: %s",
@@ -168,7 +169,8 @@ public class CmdFuck implements Command{
 
                             if(ev.getReactionEmote().getName().equals("✅")){
                                 try{
-                                    message.delete().queue();
+                                    if(message != null)
+                                        message.delete().queue();
                                 }catch(Exception ex){
                                     logger.warn(String.format(
                                             "Couldn't delete own message for CmdFuck. Reason: %s",
@@ -203,7 +205,8 @@ public class CmdFuck implements Command{
                         }, 1, TimeUnit.MINUTES,
                         () -> {
                             try {
-                                message.delete().queue();
+                                if(message != null)
+                                    message.delete().queue();
                             }catch (Exception ex){
                                 logger.warn(String.format(
                                         "Couldn't delete own message for CmdFuck. Reason: %s",
